+++ src/main/java/kitchenpos/domain/table/service/TableService.java
@@ -2,12 +2,12 @@ package kitchenpos.domain.table.service;
 
 import kitchenpos.domain.order.Order;
 import kitchenpos.domain.order.repository.OrderRepository;
-import kitchenpos.domain.order.service.dto.OrderTableCreateRequest;
-import kitchenpos.domain.order.service.dto.OrderTableResponse;
-import kitchenpos.domain.order.service.dto.OrderTableUpdateGuestsRequest;
-import kitchenpos.domain.order.service.dto.OrderTableUpdateRequest;
 import kitchenpos.domain.table.OrderTable;
 import kitchenpos.domain.table.repository.OrderTableRepository;
+import kitchenpos.domain.table.service.dto.OrderTableCreateRequest;
+import kitchenpos.domain.table.service.dto.OrderTableResponse;
+import kitchenpos.domain.table.service.dto.OrderTableUpdateGuestsRequest;
+import kitchenpos.domain.table.service.dto.OrderTableUpdateRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
