+++ src/main/java/kitchenpos/domain/table/service/dto/OrderTableUpdateGuestsRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain.order.service.dto;
+package kitchenpos.domain.table.service.dto;
 
 public class OrderTableUpdateGuestsRequest {
