+++ src/main/java/kitchenpos/domain/table/ui/TableGroupRestController.java
@@ -1,8 +1,8 @@
 package kitchenpos.domain.table.ui;
 
-import kitchenpos.domain.order.service.dto.TableGroupCreateRequest;
-import kitchenpos.domain.order.service.dto.TableGroupResponse;
 import kitchenpos.domain.table.service.TableGroupService;
+import kitchenpos.domain.table.service.dto.TableGroupCreateRequest;
+import kitchenpos.domain.table.service.dto.TableGroupResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
