+++ src/test/java/kitchenpos/application/TableGroupServiceTest.java
@@ -4,13 +4,13 @@ import kitchenpos.domain.order.Order;
 import kitchenpos.domain.order.OrderLineItems;
 import kitchenpos.domain.order.OrderStatus;
 import kitchenpos.domain.order.repository.OrderRepository;
-import kitchenpos.domain.order.service.dto.TableGroupCreateRequest;
-import kitchenpos.domain.order.service.dto.TableGroupResponse;
 import kitchenpos.domain.table.OrderTable;
 import kitchenpos.domain.table.TableGroup;
 import kitchenpos.domain.table.repository.OrderTableRepository;
 import kitchenpos.domain.table.repository.TableGroupRepository;
 import kitchenpos.domain.table.service.TableGroupService;
+import kitchenpos.domain.table.service.dto.TableGroupCreateRequest;
+import kitchenpos.domain.table.service.dto.TableGroupResponse;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
