+++ src/test/java/kitchenpos/application/TableServiceTest.java
@@ -4,13 +4,13 @@ import kitchenpos.domain.order.Order;
 import kitchenpos.domain.order.OrderLineItems;
 import kitchenpos.domain.order.OrderStatus;
 import kitchenpos.domain.order.repository.OrderRepository;
-import kitchenpos.domain.order.service.dto.OrderTableCreateRequest;
-import kitchenpos.domain.order.service.dto.OrderTableResponse;
-import kitchenpos.domain.order.service.dto.OrderTableUpdateGuestsRequest;
-import kitchenpos.domain.order.service.dto.OrderTableUpdateRequest;
 import kitchenpos.domain.table.OrderTable;
 import kitchenpos.domain.table.repository.OrderTableRepository;
 import kitchenpos.domain.table.service.TableService;
+import kitchenpos.domain.table.service.dto.OrderTableCreateRequest;
+import kitchenpos.domain.table.service.dto.OrderTableResponse;
+import kitchenpos.domain.table.service.dto.OrderTableUpdateGuestsRequest;
+import kitchenpos.domain.table.service.dto.OrderTableUpdateRequest;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
